Two major issues are currently present.
First, we need to take care of &parameter. There should be an argument on the .import_file method to deside if the &parameters are resolved or if &parameter is simply read as a string and also written as &parameter again. The commit contains a static fix to this. However the argument must be passed in order to activate or deactivate the option (currently it breaks the tests for the previous parameter handling approach but once the argument is present that should be fine). A test and some test .k files are present for this purpose pytest -k "test_deck_read_parameter_as_string" -s -v.

Second, .expand(recurse=True) method is not always working. A test and some minimal .k files are present pytest -k "test_deck_expand_multi_level" -s -v. The expand option is also tested in the test above as a combination with parameters.